This commit was generated by cvs2svn to compensate for changes in r12977, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@12978 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/findpath.c
@@ -144,7 +144,7 @@ char *findpath(const char *cmd, const char *syspath, int mode)
             int l;
 
             s = strchr(path, FINDPATH_SEPARATOR_CHAR);
-            l = s ? (s - path) : strlen(path);
+            l = s ? (s - path) : (int)strlen(path);
 
             if (l + cl > sizeof buf - 5)
                 continue;
